Before backend refactor of SecretStore
