+++ packages/gtmapi/lmsrvlabbook/api/mutations/environment.py
@@ -212,79 +212,3 @@ class StopContainer(graphene.relay.ClientIDMutation):
             cls._stop_container(lb, username)
 
         return StopContainer(environment=Environment(owner=owner, name=labbook_name))
-
-
-class InsertSecretsFile(graphene.relay.ClientIDMutation, ChunkUploadMutation):
-    class Input:
-        owner = graphene.String(required=True)
-        labbook_name = graphene.String(required=True)
-        mount_path = graphene.String(required=True)
-        chunk_upload_params = ChunkUploadInput(required=True)
-        transaction_id = graphene.String(required=True)
-
-    environment = graphene.Field(lambda: Environment)
-
-    @classmethod
-    def mutate_and_wait_for_chunks(cls, info, **kwargs):
-        return InsertSecretsFile(environment=None)
-
-    # @staticmethod
-    # def _x(labbook, ):
-    #     labbook.git.add_all()
-    #     commit = labbook.git.commit(f"adding submodule ref to link dataset {dataset_namespace}/{dataset_name}")
-    #     labbook.git.update_submodules(init=True)
-    #
-    #     # Add Activity Record
-    #     adr = ActivityDetailRecord(ActivityDetailType.DATASET, show=False, action=ActivityAction.CREATE)
-    #     adr.add_value('text/markdown',
-    #                   f"Linked Dataset `{dataset_namespace}/{dataset_name}` to "
-    #                   f"project at revision `{dataset_revision}`")
-    #     ar = ActivityRecord(ActivityType.DATASET,
-    #                         message=f"Linked Dataset {dataset_namespace}/{dataset_name} to project.",
-    #                         linked_commit=commit.hexsha,
-    #                         tags=["dataset"],
-    #                         show=True)
-    #     ar.add_detail_object(adr)
-    #     ars = ActivityStore(labbook)
-    #     ars.create_activity_record(ar)
-
-    @classmethod
-    def mutate_and_process_upload(cls, info, upload_file_path, upload_filename, **kwargs):
-        if not upload_file_path:
-            logger.error('No file uploaded')
-            raise ValueError('No file uploaded')
-
-        username = get_logged_in_username()
-        owner = kwargs.get('owner')
-        labbook_name = kwargs.get('labbook_name')
-        mount_path = kwargs.get('mount_path')
-
-        lb = InventoryManager().load_labbook(username, owner, labbook_name)
-        with lb.lock():
-            secret_store = SecretStore(lb, username)
-            inserted_path = secret_store.insert_file(upload_file_path, mount_path,
-                                                     dst_filename=upload_filename)
-
-        return InsertSecretsFile(environment=Environment(owner=owner,
-                                                         name=labbook_name))
-
-
-class RemoveSecretsFile(graphene.ClientIDMutation):
-    class Input:
-        owner = graphene.String(required=True)
-        labbook_name = graphene.String(required=True)
-        key_filename = graphene.String(required=True)
-
-    environment = graphene.Field(lambda: Environment)
-
-    @classmethod
-    def mutate_and_get_payload(cls, root, info, owner, labbook_name, key_filename, client_mutation_id=None):
-        username = get_logged_in_username()
-        lb = InventoryManager().load_labbook(username, owner, labbook_name,
-                                             author=get_logged_in_author())
-        with lb.lock():
-            secstore = SecretStore(lb, username)
-            secstore.delete_file(key_filename)
-
-        return RemoveSecretsFile(environment=Environment(owner=owner,
-                                                         name=lb.name))
